Breaking change: output directory (--outdir) is now always required
Previously the script could be called without --outdir, causing the
script to create a random output directory under /tmp. This was
useful, but it was difficult to get the directory name afterwards.
If you wish to use a random directory under /tmp, use mktemp(1) or
similar, and pass the path via --outdir.